requirements-docker.txt
# Text2ObjectSDF - full dependencies for Docker
# Base image provides: torch (CUDA), Python 3.10

# From requirements.txt
numpy
trimesh
rtree
pysdf

# Training & inference (scripts, src)
PyYAML
tensorboard
tqdm
PyMCubes
# tinycudann: installed from GitHub in Dockerfile (not on PyPI)
transformers
scipy
scikit-image
